Refactor Account id type generation

+++ src/main/java/com/bankaccount/kata/dao/AccountRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface AccountRepository {
 
     Account save(Account account);
-    Account findById(long id);
+    Account findById(Integer id);
     List<Account> findAll();
 
 }
